aider: Implemented loading `PICOVOICE_APIKEY` from environment if not explicitly passed to `WakeWordDetector` class.
# Aider chat conversation:

ASSISTANT: To implement the functionality of loading the `PICOVOICE_APIKEY` from the environment if it is not passed explicitly to the `WakeWordDetector` class, we need to modify the `WakeWordDetector_V2.py` file. Here's the change that needs to be made:

VoiceProcessingToolkit/wake_word_detector/WakeWordDetector_V2.py
<source>python
<<<<<<< SEARCH
        self.access_key = access_key
=======
        self.access_key = access_key if access_key else os.getenv('PICOVOICE_APIKEY')
>>>>>>> REPLACE
</source>

This change ensures that if `access_key` is not provided when an instance of `WakeWordDetector` is created, the class will attempt to load the `PICOVOICE_APIKEY` from the environment variables. If the environment variable is also not set, it will result in `self.access_key` being `None`, which should be handled appropriately elsewhere in the code (e.g., by raising an exception if the access key is required).