aider: Modify the `WakeWordDetector` class to trigger a function without playing the notification sound.
# Aider chat conversation:

USER: it is still not. altså to make it usable it has to be able to run a function. if we remove the notification sound and just make it trigger a function

ASSISTANT: To modify the `WakeWordDetector` class so that it triggers a function without playing the notification sound, we will adjust the `voice_loop` method in the `WakeWordDetector.py` file. We will remove the call to `self.notification_sound_manager.play()` and ensure that the `action_function` is called when the wake word is detected.

Here is the *SEARCH/REPLACE block* for the `WakeWordDetector.py` file:

VoiceProcessingToolkit/wake_word_detector/WakeWordDetector.py
<source>python
<<<<<<< SEARCH
            if self.porcupine.process(pcm) >= 0:
                if self.action_function:
                    self.action_function()
                self.notification_sound_manager.play()
                self.stop_event.set()  # Stop the loop after the wake word is detected
=======
            if self.porcupine.process(pcm) >= 0:
                if self.action_function:
                    self.action_function()
                # The notification sound line has been removed
                self.stop_event.set()  # Stop the loop after the wake word is detected
>>>>>>> REPLACE
</source>

With this change, the `voice_loop` method will call the `action_function` when the wake word is detected and then set the `stop_event` to stop the loop, without playing the notification sound. This will make the wake word detection usable for triggering a function.